gogola
add gogolas changes

+++ ConsoleApp1/ConsoleApp1/IFactorial.cs
@@ -2,6 +2,6 @@
 {
     public interface IFactorial
     {
-        int CountFactorial(int number);
+        long CountFactorial(int number);
     }
 }
